add doc link and refactor
